Move `__pip-runner__` script into a module
This makes it possible to eliminate the need for writing a copy of this
script in a temporary directory for each isolated build environment.
